created clinical-significance-select type
