+++ client/src/app/forms/config/types/clinical-significance-select/clinical-significance-select.type.html
@@ -1,0 +1,1 @@
+<p>clinical-significance-select works!</p>
