+++ client/src/app/forms/config/types/variant-origin-select/variant-origin-select.module.ts
@@ -3,9 +3,7 @@ import { CommonModule } from '@angular/common';
 import { variantOriginSelectTypeOption } from './variant-origin-select.type';
 import { ReactiveFormsModule } from '@angular/forms';
 import { FormlyModule } from '@ngx-formly/core';
-import { FormlyNzFormFieldModule } from '@ngx-formly/ng-zorro-antd/form-field';
 import { FormlySelectModule } from '@ngx-formly/core/select';
-import { NzSelectModule } from 'ng-zorro-antd/select';
 
 @NgModule({
   imports: [
@@ -13,7 +11,6 @@ import { NzSelectModule } from 'ng-zorro-antd/select';
     ReactiveFormsModule,
     FormlySelectModule,
     FormlyModule.forChild({ types: [variantOriginSelectTypeOption] }),
-    NzSelectModule,
   ]
 })
 export class CvcVariantOriginSelectTypeModule { }
