+++ client/src/app/forms/evidence-revise/evidence-revise.form.ts
@@ -25,7 +25,7 @@ import { BehaviorSubject, Subject } from 'rxjs';
 import { takeUntil } from 'rxjs/operators';
 import { $enum } from 'ts-enum-util';
 import { EvidenceItemReviseService } from './evidence-revise.service';
-import * as fieldState from './evidence-revise.state';
+import * as fieldState from '@app/forms/config/states/evidence.state';
 
 
 
@@ -218,29 +218,36 @@ export class EvidenceReviseForm implements OnInit, OnDestroy {
           },
           {
             key: 'clinicalSignificance',
-            type: 'select',
+            type: 'clinical-significance-select',
             templateOptions: {
-              label: 'Clinical Signficance',
-              helpText: 'The impact of the variant for predictive, prognostic, diagnostic, or functional evidence types. For predisposing and oncogenic evidence, impact is only applied at the assertion level and N/A should be selected here.',
-              required: true,
-              options: $enum(EvidenceClinicalSignificance)
-                .map((value, key) => {
-                  return { value: value, label: formatEvidenceEnum(value) }
-                })
-            },
-            expressionProperties: {
-              'templateOptions.options': (model: any) => {
-                return fieldState.getSignificanceOptions(model.evidenceType)
-                  .map(
-                    (value) => {
-                      return {
-                        value: value,
-                        label: formatEvidenceEnum(value)
-                      }
-                    })
-              }
+              required: true
             }
           },
+          // {
+          //   key: 'clinicalSignificance',
+          //   type: 'select',
+          //   templateOptions: {
+          //     label: 'Clinical Signficance',
+          //     helpText: 'The impact of the variant for predictive, prognostic, diagnostic, or functional evidence types. For predisposing and oncogenic evidence, impact is only applied at the assertion level and N/A should be selected here.',
+          //     required: true,
+          //     options: $enum(EvidenceClinicalSignificance)
+          //       .map((value, key) => {
+          //         return { value: value, label: formatEvidenceEnum(value) }
+          //       })
+          //   },
+          //   expressionProperties: {
+          //     'templateOptions.options': (model: any) => {
+          //       return fieldState.getSignificanceOptions(model.evidenceType)
+          //         .map(
+          //           (value) => {
+          //             return {
+          //               value: value,
+          //               label: formatEvidenceEnum(value)
+          //             }
+          //           })
+          //     }
+          //   }
+          // },
           {
             key: 'disease',
             type: 'multi-field',
